Move header menu JavaScript to menu.js
