+++ src/html/index.html
@@ -103,5 +103,6 @@
         </div>
       </div>
     </div>
+    <script src="menu.bundle.js"></script>
   </body>
 </html>
