+++ src/html/leagues.html
@@ -59,5 +59,6 @@
         </div>
       </div>
     </div>
+    <script src="menu.bundle.js"></script>
   </body>
 </html>
