+++ src/html/points.html
@@ -63,5 +63,6 @@
       </div>
     </div>
   <script src="team.bundle.js"></script>
+  <script src="menu.bundle.js"></script>
   </body>
 </html>
