+++ src/html/team.html
@@ -62,5 +62,6 @@
         </div>
       </div>
     </div>
+    <script src="menu.bundle.js"></script>
   </body>
 </html>
